created navbar component for surveys without the toggle function, fixed tutor and student view
