+++ frontend/src/components/student/studentSessionsNavbar.js
@@ -1,6 +1,3 @@
-//TODO:
-//RIGHT NOW THIS COMPONENT CONTAINS FUNCTIONALITY FOR THE PAGES, WE NEED TO SPLIT THIS UP FOR CLEAN CODE
-
 import React, { Component } from 'react';
 import MuiAppBar from '@mui/material/AppBar';
 import { styled, createTheme, ThemeProvider } from '@mui/material/styles';
@@ -16,13 +13,13 @@ import Grid from '@mui/material/Grid';
 import Link from '@mui/material/Link';
 import MenuIcon from '@mui/icons-material/Menu';
 import ChevronLeftIcon from '@mui/icons-material/ChevronLeft';
-import { mainListItems } from '../../pages/student/StudentSessionsDashboard/listSessionOptions';
+import { mainListItems } from './../../pages/student/StudentSessionsDashboard/listSessionOptions';
 import logo from '../../images/logo.png';
 import axios from 'axios';
 import LogoutIcon from '@mui/icons-material/Logout';
 import { useNavigate } from 'react-router-dom';
-import Sessions from '../../pages/student/StudentSessionsDashboard/sessions';
-import AddSession from '../../components/student/studentSessionsNavbar';
+import Sessions from './../../pages/student/StudentSessionsDashboard/sessions';
+import AddSession from './../shared/addSession';
 
 // Component to display copyright information
 function Copyright(props) {
