+++ frontend/src/pages/admin/Surveys/SurveyForm.js
@@ -3,46 +3,50 @@ import { Typography, Card, Grid, IconButton } from '@material-ui/core';
 import { ExitToApp as LogoutIcon } from '@material-ui/icons';
 import { makeStyles } from '@material-ui/core/styles';
 import { useNavigate } from 'react-router-dom';
-import logo from '../../images/logo.png';
-import Header from '../../components/shared/Header';
+import logo from '../../../images/logo.png';
+import Header from '../../../components/shared/Header';
+import PayrollNavbar from '../../../components/admin/payroll/payrollNavbar_noDropdown'; // Import the new component
 
+// ... (previous code)
 const useStyles = makeStyles((theme) => ({
   
-  root: {
-    marginTop: theme.spacing(3),
-    padding: theme.spacing(3),
-    backgroundColor: '#FFFFFF',
-    minHeight: '100vh',
-    backgroundSize: 'cover',
-    backgroundPosition: 'center',
-  },
-  header: {
-    display: 'flex',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    marginBottom: theme.spacing(3),
-  },
-  logo: {
-    maxHeight: '100px',
-    marginRight: '20px',
-    marginTop: '20px',
-  },
-  title: {
-    fontSize: '3rem',
-    fontWeight: 'bold',
-    background: '-webkit-linear-gradient(45deg, #00E1FF 30%, #00FF9F 90%)',
-    WebkitBackgroundClip: 'text',
-    WebkitTextFillColor: 'transparent',
-    textAlign: 'center',
-    marginBottom: theme.spacing(3),
-  },
-  form: {
-    marginTop: theme.spacing(2),
-    height: '80vh',
-  },
-}));
+    root: {
+      marginTop: theme.spacing(3),
+      padding: theme.spacing(3),
+      backgroundColor: '#FFFFFF',
+      minHeight: '100vh',
+      backgroundSize: 'cover',
+      backgroundPosition: 'center',
+    },
+    header: {
+      display: 'flex',
+      justifyContent: 'space-between',
+      alignItems: 'center',
+      marginBottom: theme.spacing(3),
+    },
+    logo: {
+      maxHeight: '100px',
+      marginRight: '20px',
+      marginTop: '20px',
+    },
+    title: {
+      fontSize: '3rem',
+      fontWeight: 'bold',
+      background: '-webkit-linear-gradient(45deg, #00E1FF 30%, #00FF9F 90%)',
+      WebkitBackgroundClip: 'text',
+      WebkitTextFillColor: 'transparent',
+      textAlign: 'center',
+      marginTop: theme.spacing(3),
+      marginBottom: theme.spacing(3),
+    },
+    form: {
+      marginTop: theme.spacing(2),
+      height: '80vh',
+    },
+  }));
 
 function SurveyForm() {
+  // ... (previous code)
   const classes = useStyles();
   const navigate = useNavigate();  // use useHistory hook
   const handleLogout = () => {
@@ -57,13 +61,7 @@ function SurveyForm() {
 
   return (
     <Grid className={classes.root}>
-      <Header/>
-      {/* <Grid container className={classes.header}>
-        <img onClick={navigateHomePage} className={classes.logo} src={logo} alt="Logo" />
-        <IconButton onClick={handleLogout}>
-          <LogoutIcon style={{ color: '#000000' }}/>
-        </IconButton>
-      </Grid> */}
+      <PayrollNavbar /> {/* Use the new component here */}
       <Card style={{
           backgroundColor: '#FFFFFF',
           borderRadius: '15px',
